* Rework so that all speech on demand support is in a dedicated module. * Restore 2023.3 behaviour when swapping languages
